+++ core/src/modules/DiagnosticOutputModule/ConfigOutput.cpp
@@ -213,7 +213,8 @@ void ConfigOutput::outputState(const ModelState& diagState)
      */
     Duration timeSinceOutput = meta.time() - lastOutput;
     if (timeSinceOutput.seconds() > 0
-        && (everyTS || std::fmod(timeSinceOutput.seconds(), outputPeriod.seconds()) == 0.)) {
+        && (everyTS
+            || std::fmod(timeSinceOutput.seconds(), outputPeriod.seconds()) == FloatType(0))) {
         Logged::info("ConfigOutput: Outputting " + std::to_string(state.data.size()) + " fields to "
             + currentFileName + " at " + meta.time().format() + "\n");
         meta.affixCoordinates(state);
