+++ dynamics/src/include/SlopeLimiter.hpp
@@ -112,10 +112,10 @@ public:
 
             // values of phi in the 4 nodes: lower-left, lower-right, upper-left, upper-right
             const Eigen::Vector<FloatType, 4> vertexvalues
-                = { phi(c, 0) - 0.5 * phi(c, 1) - 0.5 * phi(c, 2),
-                      phi(c, 0) + 0.5 * phi(c, 1) - 0.5 * phi(c, 2),
-                      phi(c, 0) - 0.5 * phi(c, 1) + 0.5 * phi(c, 2),
-                      phi(c, 0) + 0.5 * phi(c, 1) + 0.5 * phi(c, 2) };
+                = { phi(c, 0) - FloatType(0.5) * phi(c, 1) - FloatType(0.5) * phi(c, 2),
+                      phi(c, 0) + FloatType(0.5) * phi(c, 1) - FloatType(0.5) * phi(c, 2),
+                      phi(c, 0) - FloatType(0.5) * phi(c, 1) + FloatType(0.5) * phi(c, 2),
+                      phi(c, 0) + FloatType(0.5) * phi(c, 1) + FloatType(0.5) * phi(c, 2) };
 
             // value of phi in the midpoint
             const FloatType midvalue = phi(c, 0);
@@ -125,11 +125,13 @@ public:
                 FloatType dv = vertexvalues[i] - midvalue; // distance to midpoint
                 if (dv > 1.e-8) {
                     assert(_max(cgi + cgindices[i]) >= midvalue);
-                    al = std::min(al, std::min(1.0, (_max(cgi + cgindices[i]) - midvalue) / dv));
+                    al = std::min(
+                        al, std::min(FloatType(1), (_max(cgi + cgindices[i]) - midvalue) / dv));
                 }
                 if (dv < -1.e-8) {
                     assert(_min(cgi + cgindices[i]) <= midvalue);
-                    al = std::min(al, std::min(1.0, (_min(cgi + cgindices[i]) - midvalue) / dv));
+                    al = std::min(
+                        al, std::min(FloatType(1), (_min(cgi + cgindices[i]) - midvalue) / dv));
                 }
                 assert(al >= 0);
             }
@@ -151,10 +153,11 @@ public:
             size_t cgi = cy * (mesh.nx + 1) + cx; // index of lower-left vertex
 
             // values of (d/dx phi) in the 4 nodes: lower-left, lower-right, upper-left, upper-right
-            const Eigen::Vector<FloatType, 4> vertexvalues = {
-                phi(c, 1) - phi(c, 3) - 0.5 * phi(c, 5), phi(c, 1) + phi(c, 3) - 0.5 * phi(c, 5),
-                phi(c, 1) - phi(c, 3) + 0.5 * phi(c, 5), phi(c, 1) + phi(c, 3) + 0.5 * phi(c, 5)
-            };
+            const Eigen::Vector<FloatType, 4> vertexvalues
+                = { phi(c, 1) - phi(c, 3) - FloatType(0.5) * phi(c, 5),
+                      phi(c, 1) + phi(c, 3) - FloatType(0.5) * phi(c, 5),
+                      phi(c, 1) - phi(c, 3) + FloatType(0.5) * phi(c, 5),
+                      phi(c, 1) + phi(c, 3) + FloatType(0.5) * phi(c, 5) };
             // value of d/dx phi in the midpoint
             const FloatType midvalue = phi(c, 1);
 
@@ -163,11 +166,13 @@ public:
                 FloatType dv = vertexvalues[i] - midvalue; // distance to midpoint
                 if (dv > 1.e-8) {
                     assert(_max(cgi + cgindices[i]) >= midvalue);
-                    al = std::min(al, std::min(1.0, (_max(cgi + cgindices[i]) - midvalue) / dv));
+                    al = std::min(
+                        al, std::min(FloatType(1), (_max(cgi + cgindices[i]) - midvalue) / dv));
                 }
                 if (dv < -1.e-8) {
                     assert(_min(cgi + cgindices[i]) <= midvalue);
-                    al = std::min(al, std::min(1.0, (_min(cgi + cgindices[i]) - midvalue) / dv));
+                    al = std::min(
+                        al, std::min(FloatType(1), (_min(cgi + cgindices[i]) - midvalue) / dv));
                 }
                 assert(al >= 0);
             }
@@ -188,10 +193,11 @@ public:
             size_t cgi = cy * (mesh.nx + 1) + cx; // index of lower-left vertex
 
             // values of (d/dx phi) in the 4 nodes: lower-left, lower-right, upper-left, upper-right
-            const Eigen::Vector<FloatType, 4> vertexvalues = {
-                phi(c, 2) - phi(c, 4) - 0.5 * phi(c, 5), phi(c, 2) - phi(c, 4) + 0.5 * phi(c, 5),
-                phi(c, 2) + phi(c, 4) - 0.5 * phi(c, 5), phi(c, 2) + phi(c, 4) + 0.5 * phi(c, 5)
-            };
+            const Eigen::Vector<FloatType, 4> vertexvalues
+                = { phi(c, 2) - phi(c, 4) - FloatType(0.5) * phi(c, 5),
+                      phi(c, 2) - phi(c, 4) + FloatType(0.5) * phi(c, 5),
+                      phi(c, 2) + phi(c, 4) - FloatType(0.5) * phi(c, 5),
+                      phi(c, 2) + phi(c, 4) + FloatType(0.5) * phi(c, 5) };
             // value of d/dx phi in the midpoint
             const FloatType midvalue = phi(c, 2);
 
@@ -200,11 +206,13 @@ public:
                 FloatType dv = vertexvalues[i] - midvalue; // distance to midpoint
                 if (dv > 1.e-8) {
                     assert(_max(cgi + cgindices[i]) >= midvalue);
-                    al = std::min(al, std::min(1.0, (_max(cgi + cgindices[i]) - midvalue) / dv));
+                    al = std::min(
+                        al, std::min(FloatType(1), (_max(cgi + cgindices[i]) - midvalue) / dv));
                 }
                 if (dv < -1.e-8) {
                     assert(_min(cgi + cgindices[i]) <= midvalue);
-                    al = std::min(al, std::min(1.0, (_min(cgi + cgindices[i]) - midvalue) / dv));
+                    al = std::min(
+                        al, std::min(FloatType(1), (_min(cgi + cgindices[i]) - midvalue) / dv));
                 }
                 assert(al >= 0);
             }
