+++ dynamics/src/include/dgLimit.hpp
@@ -17,7 +17,8 @@ static void LimitMax(DGVector<3>& dg, FloatType max)
 #pragma omp parallel for
     for (long int i = 0; i < dg.rows(); ++i) {
         dg(i, 0) = std::min(max, dg(i, 0));
-        const FloatType l0 = 2.0 * std::max(fabs(dg(i, 1) + dg(i, 2)), fabs(dg(i, 1) - dg(i, 2)));
+        const FloatType l0
+            = FloatType(2.0) * std::max(fabs(dg(i, 1) + dg(i, 2)), fabs(dg(i, 1) - dg(i, 2)));
         if (l0 == 0)
             continue;
         const FloatType ex = dg(i, 0) + l0 - max;
@@ -54,7 +55,8 @@ static void LimitMin(DGVector<3>& dg, FloatType min)
 #pragma omp parallel for
     for (long int i = 0; i < dg.rows(); ++i) {
         dg(i, 0) = std::max(min, dg(i, 0));
-        const FloatType l0 = 2.0 * std::max(fabs(dg(i, 1) + dg(i, 2)), fabs(dg(i, 1) - dg(i, 2)));
+        const FloatType l0
+            = FloatType(2.0) * std::max(fabs(dg(i, 1) + dg(i, 2)), fabs(dg(i, 1) - dg(i, 2)));
         if (l0 == 0)
             continue;
