+++ dynamics/src/kokkos/KokkosBBMDynamicsKernel.cpp
@@ -86,7 +86,7 @@ void KokkosBBMDynamicsKernel<DGadvection>::updateStressHighOrderDevice(
             EdgeVec s22Gauss = s22.row(i) * PSIStress;
 
             //! Current normal stress for the evaluation of tildeP (Eqn. 1)
-            EdgeVec sigma_n = 0.5 * (s11Gauss.array() + s22Gauss.array());
+            EdgeVec sigma_n = FloatType(0.5) * (s11Gauss.array() + s22Gauss.array());
 
             //! exp(-C(1-A))
             const EdgeVec expC = (params.compactionParam * (1.0 - aGauss.array())).exp().array();
@@ -104,7 +104,7 @@ void KokkosBBMDynamicsKernel<DGadvection>::updateStressHighOrderDevice(
             // (Eqn. 7b) Select case based on sigma_n
             const auto tildeP
                 = (sigma_n.array() < 0.0)
-                      .select((-Pmax.array() / sigma_n.array()).min(1.0).matrix(), 0.);
+                      .select((-Pmax.array() / sigma_n.array()).min(1.0).matrix(), FloatType(0));
 
             // multiplicator
             const EdgeVec multiplicator
@@ -135,8 +135,8 @@ void KokkosBBMDynamicsKernel<DGadvection>::updateStressHighOrderDevice(
             s22Gauss.array() *= multiplicator.array();
             s12Gauss.array() *= multiplicator.array();
 
-            sigma_n = 0.5 * (s11Gauss.array() + s22Gauss.array());
-            const EdgeVec tau = (0.25 * (s11Gauss.array() - s22Gauss.array()).square()
+            sigma_n = FloatType(0.5) * (s11Gauss.array() + s22Gauss.array());
+            const EdgeVec tau = (FloatType(0.25) * (s11Gauss.array() - s22Gauss.array()).square()
                 + s12Gauss.array().square())
                                     .sqrt();
 
@@ -150,7 +150,7 @@ void KokkosBBMDynamicsKernel<DGadvection>::updateStressHighOrderDevice(
             // Mohr-Coulomb failure using Mssrs. Plante & Tremblay's formulation
             // sigma_s + tan_phi*sigma_n < 0 is always inside, but gives dcrit < 0
             EdgeVec dcrit
-                = (tau.array() + params.mu * sigma_n.array() > 0.)
+                = (tau.array() + params.mu * sigma_n.array() > FloatType(0))
                       .select(cohesion.array() / (tau.array() + params.mu * sigma_n.array()), 1.);
 
             // Compressive failure using Mssrs. Plante & Tremblay's formulation
